requirements.txt
streamlit==1.40.0
pandas==2.2.3
openpyxl==3.1.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1
geopy==2.4.1
oracledb==3.2.0
